requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3
newsapi-python==0.2.7
python-dateutil==2.8.2
email-validator==2.1.0.post1 